+++ sharding-spring/sharding-jdbc-spring/sharding-jdbc-spring-boot-starter/src/main/java/org/apache/shardingsphere/shardingjdbc/spring/boot/encrypt/SpringBootEncryptRuleConfigurationProperties.java
@@ -25,6 +25,6 @@ import org.springframework.boot.context.properties.ConfigurationProperties;
  *
  * @author panjuan
  */
-@ConfigurationProperties(prefix = "sharding.jdbc.config.encrypt")
+@ConfigurationProperties(prefix = "spring.shardingsphere.encrypt")
 public class SpringBootEncryptRuleConfigurationProperties extends YamlEncryptRuleConfiguration {
 }